Foenix/MCP Booting on A2560K '040

+++ src/include/gabe_reg.h
@@ -36,7 +36,7 @@
  * 16-bit register controlling the key of the A2560K built-in keyboard
  */
 
-#define GABE_MO_LEDS        ((volatile unsigned short *)0xFEC0000C)
+#define GABE_MO_LEDS        ((volatile unsigned long *)0xFEC0000C)
 #define GABE_MO_LED_0_B     0x0001      /* LED 0 (Close to DEL key) -- Blue ON */
 #define GABE_MO_LED_0_G     0x0002      /* LED 0 (Close to DEL key) -- Green ON */
 #define GABE_MO_LED_0_R     0x0004      /* LED 0 (Close to DEL key) -- Red ON */
